Submodule commit for #21.
Changes:

1. Main solution: When displaying validation message, check for the
presence of unique label id and toggle display of that particular
validation message, otherwise toggle the display by class name. Offers
backward compatibility.


2. New issue discovered: When adding a new person, an already failed
validation message is shown on the new input group if the previous focus
was on an input group that's displaying a failed error message.
Solution: when cloning a new input line, check if the current widget is
a validation message and then hide it.